fix(setup_parser_common_args): fix the --skip type
Currently, argparse's action for this argument is that of a string
value, but uses in the code treat it as a list, and expect to be able to
append to it. If you specify both -k and --skip to the `shell` command,
for example, a traceback is shown. This also means that the checks such
as `command_name in skip` are actually searching for a substring.

Alter the argument action to append, which implies a list, and clarifies
the semantics of how to skip multiple steps on the command-line.

Signed-off-by: Christopher Larson <chris.larson@siemens.com>