fix issues running the actions with 'act'
apt-get may prompt if -y isn't specified. The ln commands may fail if
the links are already present. This should only be the case if the
container isn't being reset every run, but there are arguments to change
this behavior, and using -f with ln -s is a common pattern and causes no
harm. Installing libnewt to /usr/local means that its libraries may not
be found by the python module unless we run ldconfig, so run it
explicitly.

Signed-off-by: Christopher Larson <chris.larson@siemens.com>